// src/modules/admin-users/common/permissions.enum.ts
export enum AdminPermissions {
  AWARD_TEMPLATES_ALL = 'award-templates.all',
  AWARD_TEMPLATES_CREATE = 'award-templates.create',
  AWARD_TEMPLATES_READ = 'award-templates.read',
  AWARD_TEMPLATES_UPDATE = 'award-templates.update',
  AWARD_TEMPLATES_DELETE = 'award-templates.delete',

  HABIT_TEMPLATES_ALL = 'habit-templates.all',
  HABIT_TEMPLATES_CREATE = 'habit-templates.create',
  HABIT_TEMPLATES_READ = 'habit-templates.read',
  HABIT_TEMPLATES_UPDATE = 'habit-templates.update',
  HABIT_TEMPLATES_DELETE = 'habit-templates.delete',

  HABIT_CATEGORIES_ALL = 'habit-categories.all',
  HABIT_CATEGORIES_CREATE = 'habit-categories.create',
  HABIT_CATEGORIES_READ = 'habit-categories.read',
  HABIT_CATEGORIES_UPDATE = 'habit-categories.update',
  HABIT_CATEGORIES_DELETE = 'habit-categories.delete',

  PARENTS_ALL = 'parents.all',
  PARENTS_READ = 'parents.read',
  PARENTS_UPDATE = 'parents.update',
  PARENTS_DELETE = 'parents.delete',

  CHILDREN_ALL = 'children.all',
  CHILDREN_CREATE = 'children.create',
  CHILDREN_READ = 'children.read',
  CHILDREN_UPDATE = 'children.update',
  CHILDREN_DELETE = 'children.delete',

  FAMILIES_ALL = 'families.all',
  FAMILIES_CREATE = 'families.create',
  FAMILIES_READ = 'families.read',
  FAMILIES_UPDATE = 'families.update',
  FAMILIES_DELETE = 'families.delete',
}